+++ CHANGELOG.md
@@ -279,3 +279,4 @@
 - 2024-02-04: Serde geo, follow up, mod, discovery, +6
 - 2024-02-05: Pagination mod, conf, qualify, icp, +2
 - 2024-02-05: Integ campaigns, discovery, mod, geo, +5
+- 2024-02-06: Cache task, icp, bettercontact, lead, +6